Global Styles: Add support for caption elements (#41140)
* dd caption element

* test: Revert image caption class name addition for native test fixtures

The current save output for native editor Image and Gallery blocks do
not expect this class name. I did not identify the exact cause, but this
does fix the failing existing tests.

* test: Reinstate image caption class name for Image block test fixture

The native editor test fixtures includes a Gallery block v1 (which needs
to upgraded to v2 at some point). Adding a `wp-element-caption` to this
test fixture causes parsing failure.

The lack of `wp-element-caption` class name for the Image block causes
an unnecessary block update during test runs.

The `wp-element-caption` class name was added to the native editor test
fixtures for both the Gallery and Image. The addition was only needed
for the latter. The previous commit erroneously removed the class name
for both. This commit reinstates the class name for the latter.

* update snapshot

Co-authored-by: David Calhoun <438664+dcalhoun@users.noreply.github.com>