Env: Download WordPress PHPUnit Into Container (#41780)
* Added WordPress PHPUnit Library To `wp-env`

This commit adds support for downloading the
WordPress PHPUnit source and mounting it at
/phpunit in the container. It uses the version
of WordPress installed in the container to
decide what version to download.

* Improved `WordPress/wp-develop` Git Performance

This commit minimizes the amount of data that needs
to be downloaded when we pull in the test framework.
It does this by not cloning the entire repository and
only checking out once we've set up the sparse
checkout.

* Moved test library to `/wordpress-phpunit`

This commit moves the test library from `/phpunit'
to `/wordpress-phpunit`. This change is to avoid
potential confusion. `/phpunit` sounds like a
tool while `/wordpress-phpunit` is not so
ambiguous.

* Removed `wp-phpunit/wp-phpunit` Native Support

The entire point of including the test files is that they are better
to use than a package like this. We should have an opinionated
default and allow them to override it if they really want to.

* Fixed Unit Tests

Since we've made it so that the default is no longer to support
`wp-phpunit/wp-phpunit`, we need to explicitly override the
default `wp-env` behavior. This will be cleaned up in a future
pull request that switches over entirely to the test files in the
environment.

* Documented PHPUnit Test File Inclusion

* Added `wp-phpunit/wp-phpunit` Notice to Changelog