Style engine: enqueue block supports styles in Gutenberg (#42880)
* Enqueuing block supports styles in gutenberg
Updating README.md

* get_instance() and constructor no longer needed

* remove $instance declaration

* Rename hook callback to `gutenberg_enqueue_stored_styles`
Hard code a list of core styles that we want to fetch from the store and the order in which we fetch them.

* Removing layout-block-supports store
Chaining store ids for the style tag id attribute so it's clear what's being included. Also adding a comment for debugging for the same purpose.

* Removing layout-block-supports store
Chaining store ids for the style tag id attribute so it's clear what's being included. Also adding a comment for debugging for the same purpose.
Adding tests for enqueuing styles
Updated README.md

* Assign $duplicate_selectors var to avoid duplicate implodes

* Update packages/style-engine/README.md

Co-authored-by: Andrew Serong <14988353+andrewserong@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: Andrew Serong <14988353+andrewserong@users.noreply.github.com>

* Remove unnecessary conditional

* Remove deprecation notice for now until we can deal with other usages, e.g., gallery/index.php

* For block themes, print stored styles in the header.
For classic themes, in the footer.

* Update packages/style-engine/README.md

Co-authored-by: Sören Wrede <soerenwrede@gmail.com>

Co-authored-by: Ari Stathopoulos <aristath@gmail.com>
Co-authored-by: Andrew Serong <14988353+andrewserong@users.noreply.github.com>
Co-authored-by: Sören Wrede <soerenwrede@gmail.com>