Performance tests: Infer CI environment from ENV
In #23818 we added the `--ci` command-line arg to the performance test
runner to detect if we were running in a CI context.  However, if our
script is running inside Github Actions then the `ENV['CI']` value will
be set.  In this patch we're removing the command-line arg and relying
on the `ENV` value in order to simplify the script's interface.

This is part of a larger refactor to the performance test script and is
an early change to minimize subsequent larger changes.