Performance Tests: Only run 1 round of tests during PR commits
In the performance tests CI workflow we have been running every test
suite three times for each branch under test. The goal of this work,
introduced in #33710, was to reduce variation in the reported data
from the tests.

Unfortunately after measuring the data produced by our test runs, and
by running experiments that run the test suites thirty times over, the
overall variation is explained primarily by noise in the Github Actions
container running our jobs. If running the test suites three times each
reduces the variation in the results then it's not detectable above the
amount of variation introduced beyond our control.

Because these additional rounds extend the perf-test runtime by around
twenty minutes on each PR we're reducing the number of rounds to a
single pass for PR commits. This will free up compute resources and
remove the performance tests as a bottleneck in the PR workflow.

Additional work can and should be done to further remove variance in
the testing results, but this practical step will remove an obstacle
from developer iteration speed without reducing the quality of the
numbers being reported.