Tag Processor: Add ability to stop at tag closers, if requested
`WP_HTML_Tag_Processor` introduces the ability to walk through an HTML
document and modify attributes on tag openers.

At times it could be useful to stop also at tag closers in order to
perform augmented inspection and querying of a document.

In this patch we're opening up a mode that allows just that; if
querying with the `[ "tag_closers" => "visit" ]` query parrameter
then in addition to stopping at tag openers, this will allow
stopping at all tag tokens regardless of open or close status.