Tag Processor: Prevent bugs from pre-PHP8 strspn/strcspn behavior
Follows #45537

When parsing truncated HTML it was brought to our attention that when
passing  out-of-bounds indices to `strspn()` and `strcspn()` that the
behavior is different before and after PHP8. We also realized that when
we cleaned up the problems with `substr()` we left some indices without
bounds checking and that led to a different flavor of the same problem.

When parsing the following HTML we run into warnings when calling `strspn()`
and `strcspn()`. For pre-PHP8 versions this also leads to an infinite loop
while in later versions it simply omits a warning.

```php
<!-- wp:gallery {"linkTo":"none"} -->
<figure class="wp-block-gallery has-nested-images ...
```

In this patch we're adding proper bounds checking wherever we update the
internal pointer in the Tag Processor to avoid any further out-of-bounds
issues.

While this patch fixes the core issue at stake, it's worth performing a
more complete audit of the index usage throughout the class and consider
internalizing the string methods to avoid version inconsistencies and
provide a more robust mechanism for aborting when passing the end of the
provided input document.

Props to @aidvu for quickly identifying this issue.