Remove stack exchange from new issue templates (#36158)
While questions about building blocks or core are on topic, this implies you can go there for help with anything WP related, e.g. how do make my WooCommerce cart do this, or how to I make the Astra theme do that. But this is not true, 3rd party plugin/theme dev support is off topic there. It's not a general catch all WP support forum (not even a forum).