Perf: Rely on more efficient Map for Priority Queue internal state (#47156)
The priority queue has been maintaining an array of enqueued contexts
and a `WeakMap` of enqueued callbacks for dispatching idle updates.

Priority ordering of context is maintained through the inherent order of
the array, but this comes at the cost of calling `.shift()` on every
dispatch, which rewrites the ordering of all the remaining keys or
indices of the array. When the priority queue gets large (such as in
store updates) this can incur a measurable cost in time. Additionally,
the use of the array for ordering relies on `WeakMap` semantics to
loosely hold references to the contexts, which might be removed.

In this patch we're relying solely on a single `Map`. By spec, `Map`
iteration follows the order of first insertion of its keys, which exactly
matches the needs of the priority queue iteration. We're able to avoid
using the `Array` altogether, and while not totally free, `Map` deletion
is measurably faster than `Array.prototype.shift()` and will involve at
least one fewer references to the context memory than having it in the
array _and_ the `WeakMap` did.

The goal of this patch is small and likely hard to measure. It aims
to lower memory allocation and pressure on the garbage collector during
the critical hot-patch of code which runs on every keypress or change
in the editor, so being a little better here will hopefully add up to
macro-level improvements in editor responsiveness.