[Private APIs] Make the re-registration safeguard opt-in rather than opt-out via IS_WORDPRESS_CORE (#48352)
Gutenberg introduced a system of sharing private APIs in WordPress/gutenberg#46131. One of the safeguards is a check preventing the same module from opting-in twice so that contributors cannot easily gain access by pretending to be a core module.

That safeguard is only meant for WordPress core and not for the released @wordpress packages. However, right now it is opt-out and must be explicitly disabled by developers wanting to install the @wordpress packages. Let's make it opt-out instead.

In other words:

* If we’re in WP core – prevent opting-in to private API with the same package name twice
* If we’re not in WP core – don’t prevent it 

Or:

* Before this commit, double opt-in safeguard is enabled by default
* After this commit, double opt-in safeguard is disabled by default AND WordPress core [explicitly enables it](https://github.com/WordPress/wordpress-develop/pull/4121)

The corresponding PR in `wordpress-develop` repo makes WordPress explicitly set the `IS_WORDPRESS_CORE` to true:

https://github.com/WordPress/wordpress-develop/pull/4121