Welcome Guide link points to a redirected help article (#48582)
* Link points to a redirected help article

The link title Here's a detailed guide opens in a new tab to a help article which is a redirected link. We should point directly to the correct article. My PR addresses this small change and points to https://wordpress.org/documentation/article/wordpress-block-editor/.

* Update Redirected link

* Updated redirected link

* Updated redirect link

* Reverting changes which were incorrect

Reverting this incorrect link edit. This does require changing, but will be addressed in a different PR