Fix Nav block classic menu fallback race condition (#48811)
When there are no navigation or classic menus, there is a fallback to use the page list of the navigation menu. This works well, but it isn't waiting to make sure the request for classic menus has completed before deciding if it should use the classic menu conversion process or fallback to the page list. This commit adds hasResolvedClassicMenu to the useEffect hook and early return to ensure we don't check for which menu to use until we have our classic menu request.