Block Editor Handbook: Improved the readability and phrasing and corrected mistakes (#48756)
* Improved the readability and phrasing

Corrected weird and awkward phrasings and sentences, and made some things more specific.

Some examples and thoughts behind the changes:
Why would the Block Editor documentation link to the Block Editor documentation. I'm already there right? We're linking to the How-To guides.

Or why do we write Block Inserter capitalized, but not Block Editor?

Sometimes the order of words was just plain weird. "you will find here guides to help you"

There was a quick link, just before the heading Quick Links, which didn't make sense to me. Also, it's one of the first links in the documentation and it links away from the documentation to a support article so that asks for a bit more context. Clearly stating you're going away from dev focussed content to end-user focussed content.

* Changed to sentence case