Workflows: Run PHP unit tests also against previous WP version (#46983)
Gutenberg is supposed to work on the current WordPress version, and on the previous one.

While we already have solid test coverage for the current WP version, we didn't have any coverage for the previous one. This sometimes resulted in failures when people attempted to run GB on the previous WP version, as seen e.g. #46811.

PHP unit tests should catch the bulk of issues like that (while not adding as much overhead as JS unit tests, or e2e tests).

Co-authored-by: Peter Wilson <519727+peterwilsoncc@users.noreply.github.com>