Add Port To `WP_TESTS_DOMAIN` (#49883)
* Added Port Addition/Replacement Function

This new function will allow us to add/replace ports in
certain `wp-config.php` values.

* Refactored WP_SITEURL and WP_HOME Validation

* Always Use `WP_SITEURL` When Configuring WordPress

Since `WP_SITEURL` already contains the port in both the
development and test environments, there is no reason to
try and construct the URL with the port. Additionally, it is
more accurate to use the test environment's `WP_SITEURL`
since that's what WordPress will expect due to the constant.

* Automatically Add Port To `WP_TESTS_DOMAIN`

Tests should use the `WP_TESTS_DOMAIN` constant to
construct URLs in tests. This poses a problem, however,
because the port will be part of the URL in test environments.
In my review of WordPress Core and WooCommerce, adding
the port to `WP_TESTS_DOMAIN` should not break anything
and will make the container more resilient to port customizations.

* Removed Hardcoded `localhost` From PHPUnit Tests

To add resiliency to the test suite, we should use `WP_TESTS_DOMAIN`
instead of hardcoding the domain and port in the URLs.