Mobile Release v1.93.1 (#49964)
* Release script: Update react-native-editor version to 1.93.0

* Release script: Update with changes from 'npm run core preios'

* Update Changelog

* Release script: Update react-native-editor version to 1.93.1

* Release script: Update with changes from 'npm run core preios'

* test: Expand multi-line block tests (#49732)

* test: Fix typo in Preformatted test

* test: Rich Text helpers can append text

The addition of `initialSelectionStart` and `initialSelectEnd` enable
placing the caret prior to "typing" additional text. This API was
inspired by the `testing-library/user-event` APIs.

https://testing-library.com/docs/user-event/utility#type

* test: Expand multi-line components typing tests

Tests continue typing after inserting a line break now that test helpers
support that ability.

* test: Fix List block split and merge tests

Changes to the `onChangeAndSelectText` helper means text is appended by
default. To recreate merely placing the cursor in the beginning of the
text, we must re-type the text as the `onChangeAndSelectText` changes
the text in addition to selecting due to the selection logic invoking
change logic within `RichText`.

* test: Rich text test helpers mimic user events (#49804)

* test: changeAndSelectTextOfRichText defaults cursor to end of new text

This will mimic natural typing where the cursor is placed at the end of
newly typed text by default.

* test: Consolidate multiple Rich Text change text helpers

The existing helpers contained substantial overlap and were coupled to
implementation details of the `RichText` event handlers. This refactor
creates a new, single helper to mimic typing into rich text fields as a
user might.

* test: Add selectTextInRichText helper

Simply process for selecting rich text values to improve test
comprehensibility.

* test: Simplify tests with new helpers

It is no longer necessary to pass the length of the test strings as the
new helper automatically tracks the string length to place the cursor at
the end of the new value, mimicking a user typing.

* test: Refactor selectTextInRichText

Rename helper and require range start position to improve clarity of the
purpose and function of the now `selectRangeInRichText` helper.

* test: Rename typeInRichText selection parameters

Improve clarity by further differentiating the initial and final
selection range.

* [Mobile] - Fix regression with the Color hook and ColorPanel (#49917)

* Mobile - Update ColorPanel: Move it to the global styles folder and update code to latest changes from the color hook

* Mobile - ColorSettings Palette: Avoid resetting color/gradient changes since this is now being handled by the ColorPanel component

* Mobile - CustomGradientPicker - Update code to get the correct gradient data due to the latest color hook changes

* Mobile - ColorPalette - Add accessibilityLabel with the color value

* Mobile - Segmented Control - Add optional chaining

* Mobile - Paragraph block - Add tests for color customization

* Mobile - Heading block - Add tests for color customization

* Mobile - Buttons block - Add tests for color customization

* Mobile - Integration tests - Add getGlobalStyles - To pass theme data to the editor for tests

* Mobile - Paragraph block - Test setting a theme text color and for the contrast check

* Mobile - ColorPalette - Use color name as the accessibility label

* Mobile - CustomGradientPicker - Simplify code by destructuring gradientAST when calling getGradientAstWithDefault

* Mobile - ColorPalette - Add optional chaining for allColors and gradients, in some cases these values can be undefined, depending on the theme colors/configuration. It also adds a fallback name for colors that don't have a name value

* Update Changelog

---------

Co-authored-by: David Calhoun <github@davidcalhoun.me>